feat(item-details): Add support for span data to trace-item-details
- This updates trace-item-details so it can be used to query for span
  data as well
